+++ src/File/UnprocessedFileInterface.php
@@ -1,0 +1,11 @@
+<?php
+/**
+ * @copyright 2015-2017 Contentful GmbH
+ * @license   MIT
+ */
+
+namespace Contentful\File;
+
+interface UnprocessedFileInterface extends FileInterface
+{
+}