+++ tests/E2E/UnprocessedFileInPreviewTest.php
@@ -6,8 +6,10 @@
 
 namespace Contentful\Tests\E2E;
 
+use Contentful\Link;
 use Contentful\Delivery\Client;
 use Contentful\File\UploadFile;
+use Contentful\File\LocalUploadFile;
 
 class UnprocessedFileInPreviewTest extends \PHPUnit_Framework_TestCase
 {
@@ -36,5 +38,27 @@ class UnprocessedFileInPreviewTest extends \PHPUnit_Framework_TestCase
             'https://upload.wikimedia.org/wikipedia/commons/5/5c/F_Scott_Fitzgerald_1921.jpg',
             $file->getUpload()
         );
+    }
+
+    /**
+     * Files uploaded to `https://upload.contentful.com` have an expiration date.
+     * This means that the file in this test will expire shortly after the creation of this fixture.
+     * Although the response from the Preview API won't change, it will be impossible to process the Asset
+     * using the Management API. This is irrelevant for the Delivery API, but it's good to rememember
+     * when dealing with the CMA.
+     *
+     * @vcr e2e_file_uploaded_from_unprocessed_file.json
+     */
+    public function testUploadedFromFileUnprocessed()
+    {
+        $asset = $this->client->getAsset('lp8z7n381EmisqwMgmqW2');
+
+        $file = $asset->getFile();
+
+        $this->assertInstanceOf(LocalUploadFile::class, $file);
+        $this->assertEquals('Contentful', $file->getFileName());
+        $this->assertEquals('image/svg+xml', $file->getContentType());
+        $this->assertInstanceOf(Link::class, $file->getUploadFrom());
+        $this->assertEquals('Upload', $file->getUploadFrom()->getLinkType());
     }
 }